+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.12.7
 
+- Add support for the Light Client Tracking Service
 - Update `avail-light-core` to 1.2.2
 
 ## [1.12.6](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.6) - 2025-02-10